+++ docs/scripts/views/drawer/dismissible-drawer.vue
@@ -21,12 +21,12 @@
         navId="demo-menu"
         class="demo-top-app-bar"
         v-shadow="4">
-        Persistent Drawer
+        Dismissible Drawer
       </ui-top-app-bar>
 
       <div :class="[$tt('body1'), 'demo-content']">
         <div :class="$tt('body2')">
-          <h1 :class="$tt('headline4')">Persistent Drawer</h1>
+          <h1 :class="$tt('headline4')">Dismissible Drawer</h1>
           <p>Click the menu icon above to open and close the drawer.</p>
           <ui-accordion>
             <ui-markdown :code="code[3]"></ui-markdown>
@@ -44,7 +44,7 @@ import snippets from '../../mixins/snippets';
 
 export default {
   metaInfo: {
-    titleTemplate: '%s - Persistent Drawer'
+    titleTemplate: '%s - Dismissible Drawer'
   },
   components: {
     UiDrawerCommonConntent